Cinc patches, pass #1
This commit is an initial pass at making the Cinc Server patch process
work. It successfully produces a Cinc Server with a working
`cinc-server-ctl` wrapper script.

License acceptance has been disabled.

This commit contains a temporary override to use a wordmark patched
fork of chef_backup. This workaround should be removed when that change
is merged upstream.

This is a combination of 18 commits:

* 5e08ec39d4729405088b6a60d3a6c39099dbb79c
* 112109684f5f7d1b17a564e0eb56be2db5228f04
* cedb133e8fd5f6fe4c537d7615749606c65023e0
* 50a5aae3a4db248dad899e18f76c9a78234f6bac
* 69930a4cd1298e71b6bd96ba952d7e0c295532df
* 262e1ad19b2bf400fa3357ab4181f76134b548af
* fcd46f1c22edd89927ef151a966b731dbc7aa737
* d341e07d3848b5ef326908787f0e0f22ddfd51fc
* ec6a7656be152aa279d97013d486c4b7a1282c6e
* 9e7a0a53f2f3a82ef47add800c815e0c6ed7038a
* 0997d9ec12ce89967ba13e432ace075752cd766b
* e84c0d76b3aefb00acd338c27c5702584666d5e2
* c6752a0700ef9b37b6732d05fe8febf65c1b989c
* 6a216b1be348df6b06ff8097ca1258132208e5cf
* 4b2cd1dd6e1c8be20f44e390ebe0234893493ec1
* 7f3539f293b3f29493cc94e84433bfca4bdbf5ed
* a1ed2181d8de2f9dd79929438da10d4510af2fc4
* 5128318f4b7b03cae9b89366b7a4b629868975a9

Signed-off-by: Josh Gitlin <jgitlin@pinnacle21.com>